Revert "Revert "Capture CoreCLR ServiceHub dependencies in VSIX" (#3972)"
This reverts commit d7942df09bbb770de9f9fce1caa647ee0eeb9623.
