+++ eng/Signing.props
@@ -19,6 +19,7 @@
     <FileSignInfo Include="OmniSharp.Extensions.LanguageProtocol.dll" CertificateName="3PartySHA2" />
     <FileSignInfo Include="OmniSharp.Extensions.LanguageServer.dll" CertificateName="3PartySHA2" />
     <FileSignInfo Include="OmniSharp.Extensions.LanguageServer.Shared.dll" CertificateName="3PartySHA2" />
+    <FileSignInfo Include="Nerdbank.Streams.dll" CertificateName="3PartySHA2" />
   </ItemGroup>
 
   <ItemGroup>
