+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/LegacyProjectEngineFactory_3_0.cs
@@ -4,14 +4,13 @@
 using System;
 using System.Reflection;
 using Microsoft.AspNetCore.Razor.Language;
-using Microsoft.CodeAnalysis.Razor;
 
-namespace Microsoft.VisualStudio.Editor.Razor
+namespace Microsoft.CodeAnalysis.Razor.Workspaces
 {
-    [ExportCustomProjectEngineFactory("MVC-2.0", SupportsSerialization = true)]
-    internal class LegacyProjectEngineFactory_2_0 : IProjectEngineFactory
+    [ExportCustomProjectEngineFactory("MVC-3.0", SupportsSerialization = true)]
+    internal class LegacyProjectEngineFactory_3_0 : IProjectEngineFactory
     {
-        private const string AssemblyName = "Microsoft.AspNetCore.Mvc.Razor.Extensions.Version2_X";
+        private const string AssemblyName = "Microsoft.AspNetCore.Mvc.Razor.Extensions";
         public RazorProjectEngine Create(RazorConfiguration configuration, RazorProjectFileSystem fileSystem, Action<RazorProjectEngineBuilder> configure)
         {
             // Rewrite the assembly name into a full name just like this one, but with the name of the MVC design time assembly.
@@ -25,6 +24,8 @@ namespace Microsoft.VisualStudio.Editor.Razor
 
             return RazorProjectEngine.Create(configuration, fileSystem, b =>
             {
+                CompilerFeatures.Register(b);
+
                 initializer.Initialize(b);
                 configure?.Invoke(b);
             });
