+++ src/Razor/src/Microsoft.CodeAnalysis.Remote.Razor.CoreComponents/Program.cs
@@ -1,0 +1,14 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace Microsoft.CodeAnalysis.Remote.Razor.CoreComponents
+{
+    internal static class Program
+    {
+#pragma warning disable IDE0060 // Remove unused parameter
+        public static void Main(string[] args)
+#pragma warning restore IDE0060 // Remove unused parameter
+        {
+        }
+    }
+}
