+++ src/Razor/src/Microsoft.VisualStudio.RazorExtension/ServiceHubCore/Microsoft.VisualStudio.Razor.TagHelperProvider.AssemblySearchPathsConfig.json
@@ -1,0 +1,3 @@
+﻿[
+  "..\\..\\managedLanguages\\VBCSharp\\LanguageServices\\Core"
+]
